Move test + docs
